requirements.txt
django-admin
djangorestframework
drf_yasg
django-seed